requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.10
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==23.5
djangorestframework==3.14.0
dnspython==2.6.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
pycryptodome==3.20.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.28.0
zipp==3.17.0
